Sort out Nukkit vector mess and make blocks Vector3i

+++ src/main/java/cn/nukkit/network/protocol/LevelEventPacket.java
@@ -113,9 +113,9 @@ public class LevelEventPacket extends DataPacket {
     protected void decode(ByteBuf buffer) {
         this.evid = Binary.readVarInt(buffer);
         Vector3f v = Binary.readVector3f(buffer);
-        this.x = v.x;
-        this.y = v.y;
-        this.z = v.z;
+        this.x = (float) v.x;
+        this.y = (float) v.y;
+        this.z = (float) v.z;
         this.data = Binary.readVarInt(buffer);
     }
