+++ src/main/java/cn/nukkit/network/protocol/LevelSoundEventPacket.java
@@ -275,9 +275,9 @@ public class LevelSoundEventPacket extends DataPacket {
     protected void decode(ByteBuf buffer) {
         this.event = (int) Binary.readUnsignedVarInt(buffer);
         Vector3f v = Binary.readVector3f(buffer);
-        this.x = v.x;
-        this.y = v.y;
-        this.z = v.z;
+        this.x = (float) v.x;
+        this.y = (float) v.y;
+        this.z = (float) v.z;
         this.data = Binary.readVarInt(buffer);
         this.identifier = Identifier.fromString(Binary.readString(buffer));
         this.isBabyMob = buffer.readBoolean();
