+++ src/main/java/cn/nukkit/network/protocol/RespawnPacket.java
@@ -26,9 +26,9 @@ public class RespawnPacket extends DataPacket {
     @Override
     protected void decode(ByteBuf buffer) {
         Vector3f v = Binary.readVector3f(buffer);
-        this.x = v.x;
-        this.y = v.y;
-        this.z = v.z;
+        this.x = (float) v.x;
+        this.y = (float) v.y;
+        this.z = (float) v.z;
         this.respawnState = buffer.readByte();
         this.runtimeEntityId = Binary.readEntityRuntimeId(buffer);
     }
